Make sure that include_exclude_items are quoted
Some items in the Include/Exclude lists need to be quoted to work
correctly, e.g. if you have a "Plugin" entry that contains an equals
sign.

It does not seem to make a difference if we quote everything in the
Include/Exclude list, so for simplicity I have changed the code to do
that.